+++ BambooTracker/command/pattern/set_key_on_to_step_command.hpp
@@ -7,7 +7,8 @@
 class SetKeyOnToStepCommand : public AbstractCommand
 {
 public:
-	SetKeyOnToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum, int stepNum, int noteNum, int instNum);
+	SetKeyOnToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum, int stepNum,
+						  int noteNum, bool autosetInst, int instNum);
 	void redo() override;
 	void undo() override;
 	CommandId getID() const override;
@@ -16,4 +17,5 @@ private:
 	std::weak_ptr<Module> mod_;
 	int song_, track_, order_, step_, note_, inst_;
 	int prevNote_, prevInst_;
+	bool autosetInst_;
 };